VEGA-1939 - Notify Sirius when new records created (#82)
* VEGA-1939 Add LPA applications event bridge #minor

* VEGA-1939 Capture all events on bridge #minor

* VEGA-1939 Send events to existing event bridge #minor

This should make debugging simpler, as events on the existing event
bridge are logged by ECS.

* Allow pipe to decrypt DynamoDB resources

It needs access to `kms:Decrypt` to be able to read the contents of the DynamoDB records (and therefore to pipe)

I also combined the policies into one document to make it easier to read, and sharded them by region

#patch

* Add event filter and transformation

- Only pass on the event if it was an insert
- Rearrange the payload to use our event standard rather than DynamoDB structure

#minor

* Support local environments

I say "support" but it's just making the environment buildable. Localstack [doesn't support Pipes](https://github.com/localstack/localstack/issues/7920) so we can't create that resource locally. We also can't add the IAM policy locally because it depends on KMS replica keys, which Localstack doesn't support either.

#patch

---------

Co-authored-by: Elliot Smith <elliot.smith@justice.digital.gov.uk>