Upgrade to localstack v3
Remove deprecated environment variables, and use forwarder lambda as in ministryofjustice/opg-s3-antivirus#341

This leads to some slightly awkward conditionals on the lambda so that the local versions use the forwarder and the AWS-deployed versions use container images.

Had to rename the API ID to `lpauid` to match AWS's regex expectations.

Unfortunately, this leads to a hard-fork in the Lambda: in AWS environments it uses an image and locally it needs to use the zip-based forwarder to pass events to docker. Because these are contradictory properties in Terraform, we need to optional resources. I moved them into a module to simplify a bit.

#minor